Edited admin.dashboard.manage_settings.manage_membership and admin.dashboard.manage_products.manage_producs1

+++ client_code/admin/dashboard/manage_products/manage_producs1/__init__.py
@@ -298,15 +298,15 @@ class manage_producs1(manage_producs1Template):
             return
         # Retrieve and validate input values
         try:
-            processing_fee = int(self.text_box_3.text.strip())
-            min_amount = int(self.min_amount.text.strip())
-            max_amount = int(self.max_amount.text.strip())
-            min_tenure = int(self.min_tenure.text.strip())
-            max_tenure = int(self.max_tenure.text.strip())
-            roi = int(self.text_box_5.text.strip())
-            lapsed_fee = int(self.lapsed_fee.text.strip())
-            default_fee = int(self.default_fee.text.strip())
-            npa = int(self.npa.text.strip())
+            processing_fee = float(self.text_box_3.text.strip())
+            min_amount = float(self.min_amount.text.strip())
+            max_amount = float(self.max_amount.text.strip())
+            min_tenure = float(self.min_tenure.text.strip())
+            max_tenure = float(self.max_tenure.text.strip())
+            roi = float(self.text_box_5.text.strip())
+            lapsed_fee = float(self.lapsed_fee.text.strip())
+            default_fee = float(self.default_fee.text.strip())
+            npa = float(self.npa.text.strip())
     
             # Validate non-negative values
             if processing_fee < 0 or min_amount < 0 or max_amount < 0 or min_tenure < 0 or max_tenure < 0 or roi < 0 or lapsed_fee < 0 or default_fee < 0 or npa < 0:
